Try use DataTemplates where possible
