+++ Unicord.Universal/Themes/Controls/Channels.xaml.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Unicord.Universal.Resources.Controls
+{
+    public partial class Channels
+    {
+        public Channels()
+        {
+            InitializeComponent();
+        }
+    }
+}
